feat(contracts): wire escrow lifecycle to RegistryContract (#96)
- set_registry/get_registry on ProductionEscrowContract
- On-chain record_activity + update_campaign_status + link on create
- Fail-closed when registry set but escrow not approve_contract'd
- Registry CampaignStatus adds Harvested/Failed; has_campaign_record
- Integration tests assert registry without off-chain orchestrator
- Document design and failure policy in INTEGRATION.md

Closes #96

Payout wallet: GBVHELLD2JE235Y2NGTDT3MWI3T65ON6SY4N6FBHYVDAQ5FZC2CP5QXH

+++ contracts/production_escrow/src/proptest_invariants.rs
@@ -264,7 +264,7 @@ proptest! {
         client.open_dispute(&1u64, &investor, &Symbol::new(&env, "quality"));
 
         let held: i128          = target as i128; // released=refundable=returnable=0
-        let payout_amount: i128 = (held * payout_frac_num / 100).max(1).min(held - 1);
+        let payout_amount: i128 = (held * (payout_frac_num as i128) / 100).max(1).min(held - 1);
         let expected_refundable = held - payout_amount;
 
         client.resolve_dispute(&1u64, &DisputeResolution::PartialSettlement, &payout_amount);
